Add pro forma tests
This commit adds some very basic test coverage to just verify that the
metapackage works as expected and that when we install the metapackage
the end user experience is consistent and works. It doesn't actually
test anything of substence, just that we can use the Aer simulators
and that the namespacing works correctly. This is good just so we don't
accidently make a mistake and end up breaking something moving forward.