Add script to generate authors list from git repo
This commit adds a script to clone the git repos for qiskit-terra and
qiskit-aer and pull the authors out of the git log. It clones the git
repos to a temporary directory and then runs the necessary git log
commands to pull the author information needed. It also will handle
co-authors. The git log output parsing code is based on how the PBR
python project[1] generates authors files.

Fixes: #15

[1] https://git.openstack.org/cgit/openstack-dev/pbr/