Handle duplicate names and emails
This commit adjusts the processing to try and uniquify the list. It now
does case insensitive matching for all string comparisons. It also
tracks emails and author names separately. So if a name or email has
already been added to the output list it will no longer be added again.
This has the negative effect though of whatever spelling or
capitlization for a name is encountered first it will be used,
regardless if that is the preferred. Also it should be noted that this
is still not perfect if a user has a different name spelling and a
different email address in the git log it will still be duplicated in
the output.