Update names manually
The generate authors script doesn't work 100% of the time. Sometimes the author's name in the commit isn't set to something human readable or is formatted poorly. Other times the same author uses multiple names and email addresses and ends up duplicated. This is hard to catch with the script so this commit just does these steps manually.

Co-Authored-By: mtreinish <mtreinish@kortar.org>