Overrides the command `bdist_wheel` to fail.
Just after downloading a package, if the `wheel` package is present,
`pip` creates a wheel distribution of the package. `pip` then uses
it to install the package. When a wheel distribution is used, `pip` does
not run `setup.py` preventing the reinstallation of `qiskit-terra`.

To avoid the creation of the wheel, the custom method overriding
`bdist_wheel` does nothing during the `run` method. Failing to build the
metapackage `qiskit` is intended. Since the wheel is not present, `pip`
will use the regular package and run `setup.py` as usual.