Add doc8 run to CI and lint checks (#52)
* Add doc8 run to CI and lint checks

This commit adds running doc8 to our normal CI and lint workflows. Right
now we don't have any doc verification (aside from what pylint enforces
on docstrings in the tests, which doesn't get published anywhere) and when
it comes time to publish a new release there is a scramble to fix all the
warnings (and potentially errors) in the documentation. By running doc8 we
can verify that our documentation is formatted properly for sphinx to
compile the output without issue (and without any warnings).

* Fix lint issues with aqua docs

* Fix most of the aqua docs lint issues

* Fix last lint issues

* Fix lint from rebase

* More rebase cleanups for new docs

* More rebase fixes