Pin versions and document versioning policy
This commit pins the versions of the tracked elements to ensure that we
only ever install a single version of an element when we install the
meta-package. Additionally, to make it clear why we do this, and how we
handle versioning this also adds a doc on the procedure for handling
versioning.

Fixes #141