Add DOI to generate_bibtex script (#152)
A DOI was manually added to the bibtex file in #74, however that file is
generated by a script. When we update the AUTHORS file the
generate_bibtex.py script is used to create a new up to date bibtex
file. To ensure we don't lose the DOI line from future updates this
commit adds it to the generate script.