Simplify install and uninstall stage
The install stage was going off the rails trying failing to install the
built wheel in certain scenarios. When this happened the empty namespace
package from qiskit-aer was leftover and when benchmarks ran there was
nothing leftover. To avoid this and simplify the install (and by
extension uninstall) configuration to just rely on the default working
behavior this commit removes the dependence on the qiskit-aer package.
The aer package wasn't even used directly in the benchmarks since we're
not actually running simulations (for obvious time reasons). Instead Aer
was just being used as a backend for the transpiler. But, there is no
reason we have to use Aer for that, and BasicAer works just as well.