Add asv benchmarks (#30)
* Add asv benchmarks

This commit adds an asv benchmark suite to the tests include in this
repo. This will enable us to track the performance of the overall qiskit
project over time as we make commits.

* Remove H2 and LiH dirs

* Fix most, but not all, lint

* Remove commented unused code

* Add QV circuit benchmark and fix asv build and some lint

* Make lint pass

* Purity isn't in quantum_info in 0.7

* Add qft and random circuit benchmarks

* Fix issue caused by namespace package and aer sticking around

* Make depth an independent variable for qv benchmarks

* Simplify install and uninstall stage

The install stage was going off the rails trying failing to install the
built wheel in certain scenarios. When this happened the empty namespace
package from qiskit-aer was leftover and when benchmarks ran there was
nothing leftover. To avoid this and simplify the install (and by
extension uninstall) configuration to just rely on the default working
behavior this commit removes the dependence on the qiskit-aer package.
The aer package wasn't even used directly in the benchmarks since we're
not actually running simulations (for obvious time reasons). Instead Aer
was just being used as a backend for the transpiler. But, there is no
reason we have to use Aer for that, and BasicAer works just as well.

* Clear build artifacts before install

* Fix lint failures for new pylint

* Switch back to pip for wheel builds

* Fix another lint issue

* Fix hopefully the last lint issue