Remove randomness from random_hex_circuit_benchmark
The random_hex benchmark had one np.random.randomint() call that was
used to determine if the cnots in the circuit were for odd or even
numbered qubits. While not a big difference (and difficult to detect by
visual inspection) we want to eliminate any difference between runs for
benchmarking. These subtle differences can make the results unusable and
anomaly detection more difficult. To that end this commit just sets the
random value to be fixed so the generated circuit will always be the
same. Since this changes the results the version for the benchmark is
changed so that we'll treat new results different from the previous
results with the random number.