Add documentation for the user config file (#216)
In Qiskit/qiskit-terra#2122 a new feature was added for having user
config files that let local environments override some default settings
for functions in terra. Right now this is just the circuit drawer
backend, but this will grow in the future. This commit adds
documentation on how to use this feature.