Remove aqua and chemistry as separate paths from autodoc build
Now that qiskit aqua and chemistry 0.5.0 have been released they have
their code off of the same qiskit namespace as the other packages. This
means the code in the makefile for building docs which generated
separate autodoc trees for the qiskit_aqua and qiskit_chemistry
namespaces is not needed anymore. Since when we install the qiskit
python package aqua and chemistry will be picked up by the search over
the qiskit namespace. This commit updates the makefile accordingly to
only run autodoc on the root qiskit namespace.