Switch docs jobs to run on Xenial with Python 3.7
In the process of trying to debug the floating point exception
encountered during the docs builds in #235 docs builds were run in all
our unit test environments. The python3.7 environment which runs on
xenial successfully built the docs. So instead of trying to get to the
bottom of why there is a floating point exception during the sphinx
build in the current job setup, this commit just switches to the same
node setup as the python 3.7 unit test env. This means switching to
an Ubuntu Xenial node and using Python 3.7 to run the docs builds.