Remove workaround for upgrading from qiskit < 0.6
When we switched to the qiskit package being a metapackage instead of
terra in the 0.7 release there were issues introduced by upgrading from
the pre-0.7 terra package to the metapackage that would result in a
broken install. We were able to come up with a workaround that involved
shelling out to pip to manually reinstall terra during the install
process and forcing the wheel build to fail to ensure we always
installed from sdist. However these workaround caused a certain degree
of headache for users because pip will display an error despite it not
being fatal and expected. Now that we're several releases past 0.6.x
the importance of providing a direct upgrade path is not as important
since most users will be coming from at least 0.7. The value this
provides is not as critical anymore and it's frankly outweighed now by
the user confusion. This commit removes the workarounds and documents
the manual uninstall needed for upgrading from a qiskit version < 0.7.0.

Fixes #271