Add version table doc to release notes
While we pin the versions in each qiskit meta-package release to make
determining which version is installed easier to know we don't actually
document that anywhere. This commit adds a table to the bottom of the
release notes page which will show the full table of each qiskit
component version for a given meta-package release. This will enable
users to figure out more easily which versions are installed.