Add documentation on installing aer from source
Instead of just linking to Aer's contributing guide as was done
previously. This commit adds documention on how to build in install
Aer. The build system for Aer is much more complicated than any of the
other components, with a lot of specifics for different environments.
This intends to cover the widest base and link to where you can more
information (which is not the Aer repo) if you need to figure out
specific options for building Aer.