Fix edge cases and cleanup generate_authors.py (#345)
With the addition of mailmaps (per #229) to the element repos we should
start to leverage it with all the git log commands used to collect the
authors and co-authors. While the git log commands with `--format` set
were implicitly using mailmaps already the '--use-mailmap' flag is added
to all log commands to just make it explicit. In addition a mailmap
lookup step is added for co-authors because the git log command doesn't
know how to do the transform (since it's not actually a metadata field).

At the same time this removes a duplicate copy of the function used to
check the log. It also fixes some bugs with the co-author detection that
prevented it from working.