Add the Qiskit/qiskit meta-repo to generate_authors.py
We were not tracking the contributions to the Qiskit/qiskit
meta-repository which is where we host all the documentation,
benchmarks, and anything else that is used between elements. This
commit corrects the oversight adds the Qiskit/qiskit repo to the list
of repositories to use for generating the authors list.