Added benchmarks for isometries (#315)
* Added benchmarks for isometries

* Some small fixes

* Track numerical value

* Simplify counts get logic

* Set a seed and skip faster

In an effort to reduce the noise in the measurements made by the
benchmark this commit sets a seed for all calls that take one;
transpile() and random_unitary(). It also moves the skip check to the
top of setup so we don't waste time trying to do anything if we're just
going to skip.