Fix various issues with the benchmark
This commit fixes a number of issues with the benchmark. First it
removes single entry parameters from the parameter list, there's no
reason to make them parameters if we don't use multiple values. It also
fixes a small bug in the return from the build_rb_circuit() function
where we were only returning one sequence of circuits instead of the
whole set. It also adjusts how we use random values by properly setting
a seed for both circuit building and tranpile. The last thing it does is
for consistency it pins the ignis version to a single version and sets
the benchmark version number to match that. This way if/when we need to
bump the ignis version we use in the benchmarks we'll have to bump the
benchmark's version too to match. This way we can make sure we're
using the same ignis version for all benchmark results.