Add pypi automation for releases (#365)
* Add pypi automation for releases

As part of streamlining the release process, this commit adds automation
for publishing to pypi. For the metapackage this is straightforward
since we only publish an sdist right now and there is no code we just
need to build the sdist and upload that to pypi. Once #304 lands we can
also add building a universal wheel to this job, but until we can build
a wheel there is no reason to try.

* Ensure we use supported python3