Handle missing qi modules (#368)
Qiskit/qiskit-terra#2761 removed some more of the deprecated qi modules
from terra that the state tomography benchmarks were previously relying
on. While we've been skipping these benchmarks since the tomography
module was removed from terra this caused an import error to be raised
when importing the test. To avoid the import error breaking asv runs until
we rewrite the benchmark to use ignis this commit handles an import
error and treats it as a skip condition.