Fix small typo with versions in install documentation (#373)
This commit fixes a small typo in the installation documentation about
when you need to manually upgrade from an old qiskit version. The manual
step is needed when upgrading from a version older than 0.7 to a version
newer than 0.7. This is because 0.7 is the release where the qiskit
package was split into terra and aer. The docs incorrectly had < 0.6
when it should have been < 0.7 (or <=0.6.1), which is what is listed in
the readme.