Add a note to the install from source about perils of namespaces (#374)
* Add a note to the install from source about perils of namespaces

Since the qiskit package uses namespace packaging there are a lot of
potential pitfalls when trying to mix and match element installations.
Due to the limitations in python packaging and the lack of any real
standardization there are many possible permutations. This commit
attempts to assist users by adding a note to the documentation
explaining the situation in a bit more detail. It also advises always
using a virtual environment so users can simply start from scratch when
things inevitably stop working.

* Update contributing_to_qiskit.rst

Added indentation for note