Randomize Benchmarking Performace Benchmark (#306)
* first version of benchmark/randomize_benchmarking.py

* fix lint error

* fixed a misprint

* updated benchmark test for RB

* fixed RB benchmark test following comment

* fix lint error

* Add qiskit-ignis to benchmark env

* removed nq from the parameters

* Correct setup() parameter order

* Only return circuits for transpilation

* nseeds = [1]

* Fix various issues with the benchmark

This commit fixes a number of issues with the benchmark. First it
removes single entry parameters from the parameter list, there's no
reason to make them parameters if we don't use multiple values. It also
fixes a small bug in the return from the build_rb_circuit() function
where we were only returning one sequence of circuits instead of the
whole set. It also adjusts how we use random values by properly setting
a seed for both circuit building and tranpile. The last thing it does is
for consistency it pins the ignis version to a single version and sets
the benchmark version number to match that. This way if/when we need to
bump the ignis version we use in the benchmarks we'll have to bump the
benchmark's version too to match. This way we can make sure we're
using the same ignis version for all benchmark results.

* Reduce the number of length vectors used to 1

This commit removes one of the two length vectors we were using for
constructing the RB circuits. This is done in the interest of time since
the RB benchmarking can be quite slow so this reduces the number of
input parameters we iterate over which decreases the total run time of
RB benchmarks. If in the future we improve the performance so that the
impact is minimal to the full asv run we can add back additional length
vectors for benchmarking.

* Update copyright header to use consistent format