Update state tomography benchmarks to use ignis
The state tomography benchmarks have been skipped since terra deprecated
and then removed it's support for doing tomography. The new way to run
tomography is by using ignis. This commit does just that and rewrites
the tomography benchmarks to use ignis instead of terra so that we can
now run them on the current versions of terra.