Add the Qiskit/qiskit and ibmq provider to generate_authors.py (#348)
* Add the Qiskit/qiskit meta-repo to generate_authors.py

We were not tracking the contributions to the Qiskit/qiskit
meta-repository which is where we host all the documentation,
benchmarks, and anything else that is used between elements. This
commit corrects the oversight adds the Qiskit/qiskit repo to the list
of repositories to use for generating the authors list.

* IBMQ too

* Remove chemistry from list

With https://github.com/Qiskit/qiskit-aqua/pull/592 having merged now,
the qiskit-chemistry git log has merged into the qiskit-aqua repository.
This means we no longer need to parse the log for both aqua and
chemistry since just aqua will contain all the git history.