Fix small bug in generate authors and update AUTHORS
This commit fixes a small copy and paste bug in the generate_authors.py
script when parsing co-authors for a patch. Previously the logic was
incorrectly using the mailmap if a match was found for a co-author.
Instead of using the returned value from the mailmap it would
incorrectly use the original author name from the commit message.

With the co-authors fixed and mailmaps existing in all tracked
repositories we can now update the authors list. So this updates the
author file and bibtex using the script. If any names appear incorrect
or inaccurate the mailmap file in the various repositories will need to
be updated to correct it.