Remove 6 qubit benchmark
The 6 qubit cat state tomography benchmark is a bit too slow to be
valuable. On my local desktop it's consistently more than the current
timeout of the state tomography benchmark which is 2min (the bell state
tomography does not have an duration issues, it's minimally slower with
6 qubits). We already have a ton of very slow benchmarks and the
additional value for running tomography with 6 qubits vs 5 is minimal.
So in the interest of time this commit removes the 6 qubit cases.