Fix doc8 line length lint
There was only one line that was over the line length for the docs
linter. It's normally smart enough to know that links can't be broken
and doesn't count that against the line length (unless you have text on
the same line after a link). This commit removes the disabled check and
fixes the one spot it was complaining about line length.