Fix doc version and release date
In #471 two small mistakes were made in the documentation for the
release. The first is that we neglected to bump the docs version from
the previous release, 0.11.2, to the new release version 0.12.0. The
second issue was the date in the changelog for the release was 2 days
prior to the actual release. This commit corrects both oversights and
brings everything up to date. In addition some small tweaks are made
to the release notes formatting as the adjustments for the linter made
from the reno source in terra broke bulleted lists in a couple places.