Remove unneeded line from 0.12.0 release notes
Underneath the header for the 0.12.0 release notes there was a stray
sentence leftover from the original version of #471 justifying the bump
to 0.12.0. This is not accurate because the release included a release
of every element. Additionally we don't need to justify increasing the
version number in the release notes in general, we have a documented
versioning policy and release version numbers don't really mean
anything aside from that policy. This commit removes the sentence to
clean up the release notes slightly.