Fix deploy docs script poBranch update (#508)
* Fix deploy docs script poBranch update

This commit fixes the deploy documentation script trying to push to
poBranch. Previously the the script was trying to reuse the deploy
key [1] for pushing to qiskit.org, this will not work because the deploy
key is only authorized to push to the Qiskit/qiskit.org repository. To
enable pushing updated po files on merges to master a second deploy key
is added with write permission to Qiskit/qiskit. This key is added to
the ssh agent as part of the deploy documentation script so when we push
to poBranch the user will have permission.

At the same time some of the debugging setup in this script is replaced
with turning bash trace mode on, this is easier than manually echoing
command prior to running them.

* Remove trace mode, it's a potential security risk

* Update deploy_documentation.sh

Removed redundant code and added code to git remove `/api` and `/apidocs` whenever `.po` files are updated