Swap order of keys for ssh-add in deploy_documentation
This commit swaps the order we add keys to the ssh agent in
deploy_documentation script. It is currently failing on trying to push
to poBranch with a permission denied error, which is probably because
it's using the wrong key. This changes the script to verify that it
works when the only configured key is for the respective push.