Ignore LOCALIZATION_CONTRIBUTORS from authors file (#519)
The AUTHORS file is just for tracking those who have to the qiskit
project. However, changes to the docs/LOCALIZATION_CONTRIBUTORS file
in the Qiskit/qiskit repo are not actually contributions, it's a
workaround to enable translators to sign the cla. Translations are done
through another service, crowdin, however because translators need to
sign the CLA prior to contributing and crowdin has no such mechanism they
need to create a github PR to do so. This however is not actually an
indication of actual contribution just that they signed the CLA.
Translation contributions will be tracked separately using a different
mechanism eventually, but until then we don't want to accidently include
this CLA workflow in the AUTHORS file. This commit modifies the
generate_authors file to exclude this file from the git log used to
build the AUTHORS list.