Update terra for the requirements list
The requirements list previously treated terra as a special case because
we used to have a workaround to facilitate a seamless upgrade from
Qiskit 0.6 (which became qiskit-terra) to Qiskit 0.7 (which is the
meta-package). In order to do this the terra version was stored in a
variable so it could be used in multiple places. However, this
workaround was removed in #304 but in that patch we kept the terra
version as a variable. This commit corrects the oversight and removes
the variable usage and treats the terra version in the requirements
list the same as the other elements.