Add custom sphinx directive to generate version history
This commit adds a custom sphinx directive to generate the version
history table. Hand editting the table is error prone and as the number
of versions grow it's easy to lose track of things. Especially given
that all the data needed for constructing the table is in the git
history already. After this commit instead of manually editing the table
a simple directive is called which will build the table by looking at
the git history for tagged versions of qiskit and checking the
requirements list for each tag.

Longer term this will make automating release note generation using reno
from each element much easier since we do not need to update a table
too.