Don't create the note in custom directive
Create the note nodes by hand in the custom directive is error prone and
not as a preceise as doing it via rst by calling the directive. The goal
of this PR is to make more of the releases notes file static so releases
are less error prone. Having a static note about the versioning policy
not being defined in the rst doesn't effect that goal so it's better to
just stick with that rather than try to recreate it at the docutils
level.