Remove Makefile for docs and just use tox
The Makefile useage for docs building had a number of issues. The main
issue is that it does not install any dependencies for you, it leaves
the burden up to the user to ensure they have a properly configured
python environment setup in their working environment. This can be
problematic especially given that our documentation doesn't walk you
through complete steps to successfully build the docs (qiskit needs
to be installed in addition to the sphinx requirements). Using make
also obfuscates how the docs are already built the command run is being
built by a bunch of hard coded variables and there is not a single
command that doesn't have to be constructed that people can just copy
and run locally (or set up a shell alias for)

To top it all off we're not really using the advantages of Makefile.
At the end of the day all we're doing is hard coding an alias for a
specific invocation of sphinx-build as `make doc`. It can switch the
builder with `make $builder` but this is not widely used since the
documentation is mostly consumed prebuilt via html and no one is
building local latex or applehelp versions of the docs. In addition
the one place this could have been used, for building the gettext
message catalog for translations. But because extra options are needed
there is no way to pass them in so it can't be used.

While we also use sphinx-build make mode in the makefile (which actually
doesn't need to be invoked via automake to use) this is actually not
desireable. Using make mode for sphinx build meant that we constantly
have to make sure we delete old versions of files otherwise tox sphinx
won't overwrite them. This is not desired behavior especially since the
only 2 places we build docs are locally for testing where we want it
to always rebuild or in CI for publishing where it is always a clean
environment and makes no difference (this however is easy to change
after or in this commit, by changeing -b to -M in the sphinx-build
command).

This commit replaces the automake pieces and all sphinx usage with tox.
We're already using tox for CI testing and lint checks so it's nothing
new. The primary advantage tox brings is that it adds the packaging and
environment isolation with the build process. In other words you just
run tox to build the docs it will take care of installing everything
needed but will not pollute your system's default python environment
(which is a big problem given the number of issues filed about
namespace packaging in qiskit). It does this by buildin a virtualenv
and installing all the dependencies and the meta-package itself from
the current local version in the repo into that virtualenv (if it's
already run it just reuse the venv). This gives users a reproducible way
to build docs that matches how they're actually built for production.
Just running tox -edocs will build (for english) the same as what is
hosted at: https://qiskit.org/documentation/. But it also provides us
with additional flexibility to tweak the build on demand based on what
you're doing. For example, you can build docs in parallel to speed up
the process by running tox -edocs -- -j auto. Or pass in any args to
sphinx build as needed. Additionally, tox can be run anywhere from
inside the git repo as the working directory.