Add support for building api docs from in-repo
This commit adds support for pulling in the source from each repo based
on how it's installed. If we have the metapackage (or releases of the
elements) installed this is a simple operation just pull it from git at
the installed version and copy it to the right place. However, the extra
complexity comes from supporting installs from arbitrary git versions.
We need to be able to checkout an installed copy at a particular sha1
and also copy from a local editable install. This attempts to do this
by parsing how a dev version is installed and either copying it from the
local repo (if it's an editable install) or cloning that sha1 from
github if it's a random checkout installed normally.

My personal opinion is that we we should restrict the meta-repo's docs
to only be responsible for building a combined view from releases.
Building and reviewing unreleased versions of the docs end up becoming
the domain of each individual element. This both simplifies the building
story but also reduces the complexity for translations. It becomes a
much more static target for strings to translate. The per repo docs are
mostly self contained and the only reason we need to copy them into the
meta-repo is for a uniform ToC and navigation side bar. In my ideal
workflow we'd just individually generate each repo's docs and upload the
static content to a known place on qiskit.org/documentation. But that
would take more rework of how the meta-docs are structure, and how we
link things together.