Fix locale output dir from tox gettext command
When we switched to using tox the command for generating the translation
artifacts that need to get pushed to po branch was assuming that it was
run in `docs/` but tox runs by default from the dir where tox.ini lives
(for python testing that's a sane assumption). This broke the deploy
documentation script because the locale output is no longer where it
it's expected to be (because it uses the sphinx-intl defaults instead of
the configured output location). This commit fixes the command to both
have a hardcoded output directory of docs/locale and also use the config
file docs/conf.py.