Update path for generate bibtex script (#570)
The generate_authors.py and generate_bibtex.py scripts are both expected
to be called from the repo root (so you call tools/generate_authors.py
for example) and that will update the files in your CWD. However the
generated_bibtex.py script was writing and reading the output files
assuming your CWD was the tools directory. In the near future calling
these tools will be automated as part of the release workflow and we
need to have the scripts work in the same manner. This commit updates
the generate_bibtex.py scripts so that it works the same way as
generate_authors.py.