Add building docs at tags to deploy documentation
We want to start hosting docs at releases of qiskit so that users can
switch between them. This commit adds support for ensuring we've built
and hosted docs for each version (starting with 0.8.0). This patch adds
a new function build_old_versions and calls it before we add changes to
the target repo and push it. This function loops over all the tags and
checks if we've built the docs yet or not. If we haven't it will build
the docs at that version and add it to the target repo. This will make
the first build after this merges quite slow, but future builds will
have no overhead except for when we release which will be double length
to include the tag build in addition to the regular master build.