Checkout all of master on docs builds
Since switching to building release version tables dynamically in docs
builds (and in the future having support for multiple things being
built dynamically based on version)s the version table only shows the
latest tag. This is because the git checkout from travis by default is a
shallow checkout meaning the full history isn't there. So when the
custom sphinx directive used to build the versions table runs it only
finds a single version in the checkout. To address this commit changes
the travis config for the docs job to disable the shallow checkout and
have the docs job pull the full repo. This should both fix the version
history table and also enable other git based dynamic docs generation in
the future.