Inline fake backend with noise from terra
This commit inlines the fake melbourne backend from qiskit-terra's test
module. We need a local copy of this in the benchmarks that's static for
consistency reasons. The fake provider in terra has different values
returned depending on where in the git history it is (if it exists at
all). This would cause noise in the benchmarks as we're trying to
measure the performance of the transpiler not the backends. So to avoid
those potential issues this commit adds a local copy that we can keep
static independent of terra. If we do need to change it for any reason
we can then also update the benchmark version to indicate it's different
results.