Add benchmarks for converter functions (#628)
* Add benchmarks for converter functions

This commit adds benchmarks for the converter functions. The converter
functions are normally quick to execute but are commonly used as part of
other operations. For example circuit_to_dag and dag_to_circuit get
called under the covers during transpile() calls. Individually
benchmarking the performance of them is useful so we can isolate and
track their performance, even if it's normally only a small percentage
of a larger operation.

* Pivot to use random circuit

The benchmarks now use a randomly generated circuit (with a fixed
seed for consistency) of varying sizes to test different cases. In the future
we can add additional types of circuits to try and test additional worst
case edge cases.

* Fix lint

* Fix lint for real

* Add comment about skips