Make version table extension work outside of git (#637)
This commit fixes an issue with the custom sphinx extension used to
build a version table dynamically based on the git history for the meta
repository. It was built assuming the git data from the repo will always
be present. However, in some cases (namely zip downloads from github)
the checkout isn't actually a git repo. When this happened the git
commands this extension depends on would fail breaking the build. Since
this is a possible condition this changes the extension to treat the
failure as non-fatal and instead just retur an empty result. This means
the extension won't generate any entries in the table, since there isn't
any git history to parse for version info. But, it means you can
successfully build the docs outside of a git repo.