Exclude qiskit-bot from author's list
Starting with the terra 0.9.1 release the qiskit-bot user account has
been used to run qiskit-bot [1] which is used to automate various
aspects of development and release workflows for qiskit. This includes
opening PRs to bump the meta repository at release time. Since this
leads to a commit with the author listed as qiskit-bot in the git log
when those commits are merged [2]. This has the side effect of creating
an entry in the authors list for the bot. Since it's not a real person
and not authoring a change outside of increasing versions as it's been
programmed, this commit excludes the qiskit-bot entry from the authors
list.

[1] https://github.com/Qiskit/qiskit-bot
[2] https://github.com/Qiskit/qiskit/commit/fafa867ea3051a4ab649455cfb86d8e40c5f311a