Parallelize sphinx builds for translated builds
Right now we're running the translated builds serially. This takes a lot
of time and will only get worse as we add more languages. To try and
mitigate this to a certain amount this removes the for loop in the
translated doc build with GNU parallel to run each translation build in
parallel. This should improve the speed of translation build jobs.