Add benchmark for non-mapping passes
This commit adds new benchmark methods for each non-mapping pass. There
are 2 benchmarks for each analysis pass that track the run time of the
pass and the memory consumption, for tranformation passes a third
benchmark is added to track the depth after the pass is run. The next
stage here will be to add benchmarks for mapping passes.